+++ CHANGELOG.md
@@ -31,4 +31,7 @@ Hotfixes for BaseOption flag for Renewal workflow
 Hotfix for domain lookup  
 
 1.1.2
-Hotfix for renewal workflow
+Hotfix for renewal workflow  
+
+1.2.0
+Add SyncProducts config to filter certificate sync by product ID